Move landscape and portrait versions of A4 page of captions
The single file now contains four types of pages:

  - landscape pages with 4 landscape cards
  - portrait pages with 4 portrait cards
  - landscape pages with 4 hybrid landscape or portrait cards
  - portrait pages with 4 hybrid landscape or portrait cards

Pages with hybrid cards can combine cards in portrait orientation
with cards in landscape orientation. This move was needed because
I have only 3 cards in portrait orientation in my current portfolio,
which will need to be combined with a landscape card for printing.