Remove 10mm on one side, packing for cards of width 100mm instead of 105mm
I have increased the margin by 10mm on one side and adjusted the position
of all the guide lines along that direction.

I thought previously that I would leave an empty area between the cards
in the middle of the page, but I could only set one value for the width
of gutters, the same for vertical and horizontal gutters. As a result,
it was not simple to delineate a larger space in one orientation to
account for the fact that the cards would butt up against each other
along the 297mm length of the page, but leave a 10mm gap along the
210mm width of the page. Moving the gap to the side of the page solved
the issue: the cards will now be packed tight in both directions.