Set size of all captions to 13pt for consistency
Previously, landscape cards had 15pt captions
while portrait cards had 11.5pt captions.

The scaling of 15pt to 11.5pt probably occurred
while resizing the text box through the bottom-right
corner after rotating the orientation of the page.

An intermediate size of 13pt is small enough to fit
in the width of the portrait orientation and large
enough to be read from a small distance, on the
screen or in a card held at arm length to show it.